Module ap_dataflash::messages

Structs

AccStruct

IMU accelerometer data

AdsbStruct

Automatic Dependant Serveillance - Broadcast detected vehicle information

AetrStruct

Normalised pre-mixer control surface outputs

Ahr2Struct

Backup AHRS data

Ais1Struct

Contents of ‘position report’ AIS message, see: https://gpsd.gitlab.io/gpsd/AIVDM.html#_types_1_2_and_3_position_report_class_a

Ais4Struct

Contents of ‘Base Station Report’ AIS message, see: https://gpsd.gitlab.io/gpsd/AIVDM.html#_type_4_base_station_report

Ais5Struct

Contents of ‘static and voyage related data’ AIS message, see: https://gpsd.gitlab.io/gpsd/AIVDM.html#_type_5_static_and_voyage_related_data

AisrStruct

Raw AIS AVDIM messages contents, see: https://gpsd.gitlab.io/gpsd/AIVDM.html#_aivdmaivdo_sentence_layer

AoaStruct

Angle of attack and Side Slip Angle values

ArmStruct

Arming status changes

ArspStruct

Airspeed sensor data

Asm1Struct

AirSim simulation data

Asm2Struct

More AirSim simulation data

AtdeStruct

AutoTune data packet

AtrpStruct

Plane AutoTune

AttStruct

Canonical vehicle attitude

AtunStruct

Copter/QuadPlane AutoTune

AuxfStruct

Auxiliary function invocation information

BaroStruct

Gathered Barometer data

BatStruct

Gathered battery data

Bcl2Struct

Battery cell voltage information

BclStruct

Battery cell voltage information

BcnStruct

Beacon information

CamStruct

Camera shutter information

CmdStruct

Executed mission command information

CmdiStruct

Generic CommandInt message logger(CMDI)

CsrvStruct

Servo feedback data

CtrlStruct

Attitude Control oscillation monitor diagnostics

CtunStruct

Control Tuning information

DmsStruct

DataFlash-Over-MAVLink statistics

DsfStruct

Onboard logging statistics

DstlStruct

Deepstall Landing data

Eah1Struct

External AHRS data

EcylStruct

EFI per-cylinder information

Efi2Struct

Electronic Fuel Injection system data - redux

EfiStruct

Electronic Fuel Injection system data

ErrStruct

Specifically coded error messages

EscStruct

Feedback received from ESCs

EvStruct

Specifically coded event messages

FmtStruct

Message defining the format of messages in this file

FmtuStruct

Message defining units and multipliers used for fields of other messages

FollStruct

Follow library diagnostic data

Ftn1Struct

FFT Filter Tuning

Ftn2Struct

FFT Noise Frequency Peak

FtnStruct

Filter Tuning Messages

GpaStruct

GPS accuracy information

GpsStruct

Information received from GNSS systems attached to the autopilot

GrawStruct

Raw uBlox data

GrxhStruct

Raw uBlox data - header

GrxsStruct

Raw uBlox data - space-vehicle data

GyrStruct

IMU gyroscope data

HeatStruct

IMU Heater data

IcmbStruct

ICM20789 diagnostics

ImuStruct

Inertial Measurement Unit data

IomcStruct

IOMCU diagnostic information

Jsn1Struct

Log data received from JSON simulator

Jsn2Struct

Log data received from JSON simulator

LandStruct

Slope Landing data

LgrStruct

Landing gear information

MagStruct

Information received from compasses

MavStruct

GCS MAVLink link statistics

MavcStruct

MAVLink command we have just executed

MmoStruct

MMC3416 compass data

ModeStruct

vehicle control mode information

MonStruct

Main loop stuck data

MotbStruct

Motor mixer information

MsgStruct

Textual messages

MultStruct

Message mapping from single character to numeric multiplier

Nkf0Struct

EKF2 beacon sensor diagnostics

Nkf1Struct

EKF2 estimator outputs

Nkf2Struct

EKF2 estimator secondary outputs

Nkf3Struct

EKF2 innovations

Nkf4Struct

EKF2 variances SV, SP, SH and SM are probably best described as ‘Squared Innovation Test Ratios’ where values <1 tells us the measurement was accepted and >1 tells us it was rejected. They represent the square of the (innovation / maximum allowed innovation) where the innovation is the difference between predicted and measured value and the maximum allowed innovation is determined from the uncertainty of the measurement, uncertainty of the prediction and scaled using the number of standard deviations set by the innovation gate parameter for that measurement, eg EK2_MAG_I_GATE, EK2_HGT_I_GATE, etc

Nkf5Struct

EKF2 Sensor innovations (primary core) and general dumping ground

NkqStruct

EKF2 quaternion defining the rotation from NED to XYZ (autopilot) axes

NktStruct

EKF2 timing information

Nky0Struct

EKF Yaw Estimator States

Nky1Struct

EKF Yaw Estimator Innovations

NtunStruct

Navigation Tuning information - e.g. vehicle destination

OabrStruct

Object avoidance (Bendy Ruler) diagnostics

OadjStruct

Object avoidance (Dijkstra) diagnostics

OavgStruct

Object avoidance path planning visgraph points

OfStruct

Optical flow sensor data

OfcaStruct

Optical Flow Calibration sample

OfgStruct

OFfboard-Guided - an advanced version of GUIDED for companion computers that includes rate/s.

OrgnStruct

Vehicle navigation origin or other notable position

ParmStruct

parameter value

PidaStruct

Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Altitude/Steering

PidgStruct

Plane Proportional/Integral/Derivative gain values for Heading when using COMMAND_INT control.

PidpStruct

Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Altitude/Steering

PidrStruct

Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Altitude/Steering

PidsStruct

Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Altitude/Steering

PidyStruct

Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Altitude/Steering

PiqaStruct

QuadPlane Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Z

PiqpStruct

QuadPlane Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Z

PiqrStruct

QuadPlane Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Z

PiqyStruct

QuadPlane Proportional/Integral/Derivative gain values for Roll/Pitch/Yaw/Z

PlStruct

Precision Landing messages

PmStruct

autopilot system performance and general data dumping ground

PosStruct

Canonical vehicle position

PowrStruct

System power information

PrtnStruct

Plane Parameter Tuning data

PrxStruct

Proximity Filtered sensor data

PrxrStruct

Proximity Raw sensor data

PscdStruct

Position Control Down

PsceStruct

Position Control East

PscnStruct

Position Control North

QtunStruct

QuadPlane vertical tuning message

RadStruct

Telemetry radio statistics

RalyStruct

Rally point information

RateStruct

Desired and achieved vehicle attitude rates. Not logged in Fixed Wing Plane modes.

RbchStruct

Replay Data Beacon Header

RbciStruct

Replay Data Beacon Instance

RbohStruct

Replay body odometry data

RbrhStruct

Replay Data Barometer Header

RbriStruct

Replay Data Barometer Instance

RcdaStruct

Raw RC data

Rci2Struct

(More) RC input channels to vehicle

RcinStruct

RC input channels to vehicle

RcouStruct

Servo channel output values

RephStruct

Replay external position data

Rev2Struct

Replay Event

RevhStruct

Replay external position data

Rey3Struct

Replay Euler Yaw event

RfndStruct

Rangefinder sensor information

RgphStruct

Replay Data GPS Header

RgpiStruct

Replay Data GPS Instance, infrequently changing data

RgpjStruct

Replay Data GPS Instance - rapidly changing data

RmghStruct

Replay Data Magnetometer Header

RmgiStruct

Replay Data Magnetometer Instance

RofhStruct

Replay optical flow data

RpmStruct

Data from RPM sensors

RrnhStruct

Replay Data Rangefinder Header

RrniStruct

Replay Data Rangefinder Instance

Rso2Struct

Replay Set Origin event

RssiStruct

Received Signal Strength Indicator for RC receiver

RvohStruct

Replay Data Visual Odometry data

Rwa2Struct

Replay set-default-airspeed event

RwohStruct

Replay wheel odometry data

SaStruct

Simple Avoidance messages

SbphStruct

Swift Health Data

SbrhStruct

Swift Raw Message Data

ScrStruct

Scripting runtime stats

Sim2Struct

Additional simulator state

SimStruct

SITL simulator state

SitlStruct

Simulation data

Slv1Struct

Log data received from JSON simulator 1

Slv2Struct

Log data received from JSON simulator 2

SmooStruct

Smoothed sensor data fed to EKF to avoid inconsistencies

SoarStruct

Logged data from soaring feature

SrtlStruct

SmartRTL statistics

StakStruct

Stack information

StatStruct

Current status of the aircraft

Tec2Struct

Additional Information about the Total Energy Control System

TecsStruct

Information about the Total Energy Control System

TerrStruct

Terrain database infomration

TrigStruct

Camera shutter information

TrmpStruct

Torqeedo Motor Param

TrmsStruct

Torqeedo Motor Status

TrqdStruct

Torqeedo Status

TrseStruct

Torqeedo System Setup

TrstStruct

Torqeedo System State

TsynStruct

Time synchronisation response information

Ubx1Struct

uBlox-specific GPS information (part 1)

Ubx2Struct

uBlox-specific GPS information (part 2)

UnitStruct

Message mapping from single character to SI unit

VarStruct

Variometer data

VibeStruct

Processed (acceleration) vibration information

VisoStruct

Visual Odometry

VispStruct

Vision Position

VisvStruct

Vision Velocity

VstbStruct

Log message for video stabilisation software such as Gyroflow

WencStruct

Wheel encoder measurements

WincStruct

Winch

WindStruct

Windvane readings

Xkf0Struct

EKF3 beacon sensor diagnostics

Xkf1Struct

EKF3 estimator outputs

Xkf2Struct

EKF3 estimator secondary outputs

Xkf3Struct

EKF3 innovations

Xkf4Struct

EKF3 variances. SV, SP, SH and SM are probably best described as ‘Squared Innovation Test Ratios’ where values <1 tells us the measurement was accepted and >1 tells us it was rejected. They represent the square of the (innovation / maximum allowed innovation) where the innovation is the difference between predicted and measured value and the maximum allowed innovation is determined from the uncertainty of the measurement, uncertainty of the prediction and scaled using the number of standard deviations set by the innovation gate parameter for that measurement, eg EK3_MAG_I_GATE, EK3_HGT_I_GATE, etc

Xkf5Struct

EKF3 Sensor innovations (primary core) and general dumping ground

XkfdStruct

EKF3 Body Frame Odometry errors

XkfmStruct

EKF3 diagnostic data for on-ground-and-not-moving check

XkfsStruct

EKF3 sensor selection

XkqStruct

EKF3 quaternion defining the rotation from NED to XYZ (autopilot) axes

XktStruct

EKF3 timing information

XktvStruct

EKF3 Yaw Estimator States

Xkv1Struct

EKF3 State variances (primary core)

Xkv2Struct

more EKF3 State Variances (primary core)

Xky0Struct

EKF Yaw Estimator States

Xky1Struct

EKF Yaw Estimator Innovations

Enums

LogType

Messages